feat(tests): add new macronizer tests and remove old tests from tests.rs
